+++ etl/scripts-ccao-data-raw-us-east-1/housing/housing-ari_index.py
@@ -1,26 +1,80 @@
-# type: ignore
-# pylint: skip-file
-# sc.addPyFile(  # noqa: F821
-#     "s3://ccao-dbt-athena-dev-us-east-1/packages/spark-packages.zip"
-# )
-
 import os
-import tempfile
 
 import boto3
-import pandas as pd
+import requests
+from botocore.exceptions import NoCredentialsError
+from bs4 import BeautifulSoup
 
-# Set up the S3 client
-s3 = boto3.client("s3")
-temp_file = tempfile.NamedTemporaryFile(suffix=".xlsx", delete=False)
 
-# Download the file from S3 to your local system
-AWS_S3_RAW_BUCKET = "s3://ccao-data-raw-us-east-1/"
-file_key = os.path.join("housing", "ari_index", "2023-ARI.xlsx")
+def get_most_recent_ihs_data_url(base_url):
+    headers = {
+        "User-Agent": (
+            "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 "
+            "(KHTML, like Gecko) Chrome/58.0.3029.110 Safari/537.3"
+        )
+    }
+    response = requests.get(base_url, headers=headers)
+    response.raise_for_status()
 
-s3.download_file(AWS_S3_RAW_BUCKET, file_key, temp_file.name)
+    soup = BeautifulSoup(response.text, "html.parser")
+    links = soup.find_all("a", href=True)
 
-# Use pandas to read the Excel file, skipping the first two rows
-data = pd.read_excel(temp_file.name, skiprows=2, engine="openpyxl")
+    xlsx_links = [
+        link["href"] for link in links if link["href"].endswith(".xlsx")
+    ]
+    if xlsx_links:
+        most_recent_xlsx = xlsx_links[0]
+        if not most_recent_xlsx.startswith("http"):
+            most_recent_xlsx = requests.compat.urljoin(
+                base_url, most_recent_xlsx
+            )
+        return most_recent_xlsx
+    else:
+        return None
 
-temp_file.close()
+
+def download_and_upload_excel(file_url, bucket, key_prefix):
+    try:
+        headers = {
+            "User-Agent": (
+                "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 "
+                "(KHTML, like Gecko) Chrome/58.0.3029.110 Safari/537.3"
+            )
+        }
+        response = requests.get(file_url, headers=headers)
+        response.raise_for_status()
+
+        # Initialize a session using boto3
+        session = boto3.Session()
+        s3 = session.resource("s3")
+
+        # Upload directly from memory
+        object_key = f"{key_prefix}/{file_url.split('/')[-1]}"
+        s3.Bucket(bucket).put_object(Key=object_key, Body=response.content)
+
+        return f"File uploaded to {bucket}/{object_key}"
+    except requests.exceptions.HTTPError as e:
+        return f"Failed to download the file: {e}"
+    except NoCredentialsError:
+        return "Credentials not available for AWS S3"
+    except Exception as e:
+        return f"An error occurred: {e}"
+
+
+# URL of the webpage to scan for Excel files
+base_url = (
+    "https://www.ihda.org/developers/market-research/affordability-risk-index/"
+)
+most_recent_file_url = get_most_recent_ihs_data_url(base_url)
+
+# S3 Bucket and Key Prefix
+AWS_S3_RAW_BUCKET = os.environ.get("AWS_S3_RAW_BUCKET")
+key_prefix = os.path.join("housing", "ari_index")
+
+if most_recent_file_url:
+    result = download_and_upload_excel(
+        most_recent_file_url, AWS_S3_RAW_BUCKET, key_prefix
+    )
+    print(result)
+else:
+    print("No .xlsx file found on the page.")
